Address checkstyle and pmd
