+++ src/integrationTest/java/uk/gov/hmcts/darts/event/service/impl/EventProcessorTest.java
@@ -3,28 +3,19 @@ package uk.gov.hmcts.darts.event.service.impl;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.transaction.annotation.Transactional;
 import uk.gov.hmcts.darts.arm.component.AutomatedTaskProcessorFactory;
 import uk.gov.hmcts.darts.common.entity.EventEntity;
-import uk.gov.hmcts.darts.common.entity.HearingEntity;
 import uk.gov.hmcts.darts.common.repository.EventRepository;
-import uk.gov.hmcts.darts.common.util.DateConverterUtil;
-import uk.gov.hmcts.darts.event.service.EventProcessor;
 import uk.gov.hmcts.darts.testutils.PostgresIntegrationBase;
 import uk.gov.hmcts.darts.testutils.stubs.EventStub;
 import uk.gov.hmcts.darts.testutils.stubs.HearingStub;
 
-import java.time.OffsetDateTime;
-import java.time.ZoneOffset;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-public class EventProcessorTest extends PostgresIntegrationBase {
+class EventProcessorTest extends PostgresIntegrationBase {
 
     @Autowired
     private EventRepository eventRepository;
@@ -37,10 +28,9 @@ public class EventProcessorTest extends PostgresIntegrationBase {
 
     @Autowired
     private AutomatedTaskProcessorFactory eventProcessorFactory;
-    private static final OffsetDateTime STARTED_AT = OffsetDateTime.of(2024, 10, 10, 10, 0, 0, 0, ZoneOffset.UTC);
 
     @Test
-    public void testProcess() {
+    void testProcess() {
         dartsDatabase.createCase("Bristol", "case1");
         dartsDatabase.createCase("Bristol", "case2");
         dartsDatabase.createCase("Bristol", "case3");
@@ -52,7 +42,7 @@ public class EventProcessorTest extends PostgresIntegrationBase {
 
         // assert that only one of the event ids is set to current
         for (Integer eventId  : processedCurrentEventIds) {
-            Assertions.assertTrue(isOnlyOneOfTheEventIdSetToCurrent(eventIdMap.get(eventId)));
+            Assertions.assertTrue(eventStub.isOnlyOneOfTheEventIdSetToCurrent(eventIdMap.get(eventId)));
             Assertions.assertTrue(eventIdMap.get(eventId).get(1).getIsCurrent());
         }
 
@@ -68,26 +58,11 @@ public class EventProcessorTest extends PostgresIntegrationBase {
 
         // assert that only one of the event ids is set to current
         for (Integer eventId  : processedCurrentEventIds2) {
-            Assertions.assertTrue(isOnlyOneOfTheEventIdSetToCurrent(eventIdMap.get(eventId)));
+            Assertions.assertTrue(eventStub.isOnlyOneOfTheEventIdSetToCurrent(eventIdMap.get(eventId)));
             Assertions.assertTrue(eventIdMap.get(eventId).get(1).getIsCurrent());
         }
 
         // process third batch which is expected to be empty
         Assertions.assertTrue(eventProcessorFactory.createEventProcessor(1).processCurrentEvent().isEmpty());
-    }
-
-    private boolean isOnlyOneOfTheEventIdSetToCurrent(List<EventEntity> eventEntityList) {
-        boolean currentFnd= false;
-        for (EventEntity event : eventEntityList) {
-            Optional<EventEntity> readEventEntity = eventRepository.findById(event.getId());
-
-            if (readEventEntity.isPresent() && readEventEntity.get().getIsCurrent() && !currentFnd) {
-                currentFnd = true;
-            } else if (readEventEntity.isPresent() && readEventEntity.get().getIsCurrent()) {
-                return false;
-            }
-        }
-
-        return currentFnd;
     }
 }
